JWT authentication

+++ controllers/api/v1/posts_api.js
@@ -8,7 +8,7 @@ module.exports.index = async function(req,res){
         .populate({
             path:'comments',
             populate:{
-                path:'user'
+            path:'user'
         }
     });
     return res.json(200,{
@@ -21,39 +21,29 @@ module.exports.index = async function(req,res){
 module.exports.destroy =async function(req,res){
     try{
         let post = await Post.findById(req.params.id);
-        
-        // if(post.user == req.user.id){
+
+        if(post.user == req.user.id){
             post.remove();
             
             await Comment.deleteMany({post:req.param.id});
             
-            // if(req.xhr){
-            //     return res.status(200).json({
-            //         data:{
-            //             post_id:req.params.id
-            //         },
-            //         message:"Post deleted"
-            //     });
-            // }
+          
 
             return res.json(200,{
                 message:"post deleted successfully"
                 
-            })
+            });
+        }else{
+            returnres.json(401 ,{
+                message:"You can't delete this post!"
+            });
+        }
 
-        // }else{
-        //     req.flash('success','You can\'t delete this post');
-
-        //     return res.redirect('back');
-
-        // }
+       
         
         } catch(err){
             console.log('error',err);
-            
-            
-
-            return res.json(200,{
+            return res.json(500,{
                 message:"internal server error"
             })
         }
